rename the group id
Rename from ugly syyllab to dsngroup.
