+++ broke/src/main/java/org/dsngroup/broke/protocol/package-info.java
@@ -15,7 +15,10 @@
  */
 
 /**
- * The main package used for broke.
+ * This package used for dealing with protocol,
+ * e.g. qos settings, consumer group, critical options, etc.
+ * The final result is to customized an protocol header (send),
+ * and parse the protocol header (receive) to call the associate callback.
  */
 
-package org.syyllab.broke;
+package org.dsngroup.broke.protocol;
